requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.2.1
Automat==20.2.0
boto3==1.23.10
botocore==1.26.10
certifi==2022.6.15
cffi==1.15.1
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
Django==3.2.13
django-ckeditor==5.7.0
django-crispy-forms==1.13.0
django-filter==21.1
django-js-asset==2.0.0
django-recaptcha==3.0.0
django-storages==1.12.3
django-tables2==2.4.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==0.10.0
msgpack==1.0.4
Pillow==8.2.0
pkg_resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tenacity==8.0.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.9
zope.interface==5.4.0
